requirements.txt
fastapi
uvicorn
jinja2
pyyaml
python-multipart
openai